refactor: remove `_` from `_hooks`
