+++ src/site/_collections/hooks/authors.js
