+++ src/site/content/en/authors/feed.11tydata.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-const {feed} = require('../../../_collections/_hooks/authors');
+const {feed} = require('../../../_collections/hooks/authors');
 
 module.exports = {
   pagination: {
